+++ packages/preact/src/components/FieldArray.tsx
@@ -11,6 +11,7 @@ import type {
   MaybeArray,
   ResponseData,
   ValidateFieldArray,
+  ValidationMode,
 } from '../types';
 import { initializeFieldArrayStore } from '../utils';
 
@@ -43,6 +44,8 @@ export type FieldArrayProps<
     store: FieldArrayStore<TFieldValues, TFieldArrayName>
   ) => ComponentChild;
   validate?: Maybe<MaybeArray<ValidateFieldArray<number[]>>>;
+  validateOn?: Maybe<ValidationMode>;
+  revalidateOn?: Maybe<ValidationMode>;
   keepActive?: Maybe<boolean>;
   keepState?: Maybe<boolean>;
 };