+++ packages/preact/src/hooks/useLifecycle.ts
@@ -15,6 +15,7 @@ import type {
   TransformField,
   ValidateField,
   ValidateFieldArray,
+  ValidationMode,
 } from '../types';
 import { getUniqueId, updateFormState } from '../utils';
 
@@ -33,6 +34,8 @@ type LifecycleProps<
     | MaybeArray<ValidateField<FieldPathValue<TFieldValues, TFieldName>>>
     | MaybeArray<ValidateFieldArray<number[]>>
   >;
+  validateOn?: Maybe<ValidationMode>;
+  revalidateOn?: Maybe<ValidationMode>;
   transform?: Maybe<
     MaybeArray<TransformField<FieldPathValue<TFieldValues, TFieldName>>>
   >;
@@ -56,6 +59,8 @@ export function useLifecycle({
   name,
   store,
   validate,
+  validateOn,
+  revalidateOn,
   transform,
   keepActive = false,
   keepState = true,
@@ -68,6 +73,10 @@ export function useLifecycle({
         : [validate]
       : [];
 
+    // Set validation mode overrides
+    store.validateOn = validateOn;
+    store.revalidateOn = revalidateOn;
+
     // Add transformation functions
     if ('transform' in store) {
       store.transform = transform
@@ -76,7 +85,7 @@ export function useLifecycle({
           : [transform]
         : [];
     }
-  }, [store, transform, validate]);
+  }, [store, transform, validate, validateOn, revalidateOn]);
 
   useEffect(() => {
     // Create unique consumer ID