+++ packages/react/src/types/field.ts
@@ -2,6 +2,7 @@ import type { Signal } from '@preact/signals-react';
 import type { SyntheticEvent } from 'react';
 import type { FieldPath, FieldPathValue } from './path';
 import type { MaybeValue, MaybePromise, Maybe } from './utils';
+import type { ValidationMode } from './form';
 
 /**
  * Value type of the field value.
@@ -90,6 +91,8 @@ export type InternalFieldStore<
 
   // Other
   validate: ValidateField<FieldPathValue<TFieldValues, TFieldName>>[];
+  validateOn: Maybe<ValidationMode>;
+  revalidateOn: Maybe<ValidationMode>;
   transform: TransformField<FieldPathValue<TFieldValues, TFieldName>>[];
   consumers: Set<number>;
 };