+++ packages/solid/src/utils/initializeFieldStore.ts
@@ -66,6 +66,8 @@ export function initializeFieldStore(
 
       // Other
       validate: [],
+      validateOn: undefined,
+      revalidateOn: undefined,
       transform: [],
       consumers: new Set(),
     };